Moved logic on whether to update a particular Kubernetes resource into the Kubernetes deployer
The main module where the reconciliation used to happen now simply tells the deployer to ensure that a particular configuration is set. It is up to the deployer to do reads/writes against its backend as necessary.

This now makes it possible to check the actual contents of each Kubernetes object, not just the revision.